functions and nested loop - advanced

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,7 +5,7 @@
  */
 int main(void)
 {
-	int i, l, k = 1, j = 1;
+	long int i, l, k = 1, j = 1;
 
 	for (i = 1; i <= 50; i++)
 	{
